Created suit validator.

+++ src/main/java/com/epam/horsebetting/receiver/impl/SuitReceiverImpl.java
@@ -8,6 +8,7 @@ import com.epam.horsebetting.exception.ReceiverException;
 import com.epam.horsebetting.receiver.AbstractReceiver;
 import com.epam.horsebetting.receiver.SuitReceiver;
 import com.epam.horsebetting.request.RequestContent;
+import com.epam.horsebetting.validator.SuitValidator;
 import org.apache.logging.log4j.Level;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
@@ -31,27 +32,35 @@ public class SuitReceiverImpl extends AbstractReceiver implements SuitReceiver {
      */
     @Override
     public void ajaxObtainSuitsList(RequestContent content) throws ReceiverException {
-        int page = 1;
         String pageNumber = content.findParameter("page");
+        SuitValidator validator = new SuitValidator();
 
-        //TODO create validators
-        if (pageNumber != null) {
-            try {
-                page = Integer.parseInt(pageNumber);
-            } catch (NumberFormatException e) {
-                throw new ReceiverException("Page incorrect: " + e.getMessage(), e);
+        ArrayList<String> errors;
+
+        if (validator.validatePage(pageNumber)) {
+            final int step = 10;
+            final int page = Integer.parseInt(pageNumber);
+            final int offset = step * (page - 1);
+
+            try (SuitDAOImpl suitDAO = new SuitDAOImpl(false)) {
+                List<Suit> suits = suitDAO.obtainPart(step, offset);
+
+                content.insertJsonAttribute("success", true);
+                content.insertJsonAttribute("suits", suits);
+
+                LOGGER.log(Level.DEBUG, "Obtained suits part: " + Arrays.toString(suits.toArray()));
+            } catch (DAOException e) {
+                errors = new ArrayList<>();
+                errors.add("Something went wrong...");
+
+                content.insertJsonAttribute("success", false);
+                content.insertJsonAttribute("errors", errors);
+
+                throw new ReceiverException("Database Error: " + e.getMessage(), e);
             }
-        }
-
-        final int step = 10;
-        final int offset = step * (page - 1);
-
-        try (SuitDAOImpl suitDAO = new SuitDAOImpl(false)) {
-            List<Suit> suits = suitDAO.obtainPart(step, offset);
-            LOGGER.log(Level.DEBUG, "Suits part: " + Arrays.toString(suits.toArray()));
-            content.insertJsonAttribute("suits", suits);
-        } catch (DAOException e) {
-            throw new ReceiverException("Database Error: " + e.getMessage(), e);
+        } else {
+            content.insertJsonAttribute("success", false);
+            content.insertJsonAttribute("errors", validator.getErrors());
         }
     }
 
@@ -63,35 +72,45 @@ public class SuitReceiverImpl extends AbstractReceiver implements SuitReceiver {
      */
     @Override
     public void createSuit(RequestContent content) throws ReceiverException {
-        String name = content.findParameter("name");
+        SuitValidator validator = new SuitValidator();
         ArrayList<String> messages = new ArrayList<>();
 
-        //TODO Create validators
+        String name = content.findParameter("name");
 
-        Suit suit = new Suit(name);
-        LOGGER.log(Level.DEBUG, "Want create suit: " + suit);
+        if (validator.validateName(name)) {
+            Suit suit = new Suit(name);
+            LOGGER.log(Level.DEBUG, "Want create suit: " + suit);
 
-        SuitDAOImpl suitDAO = new SuitDAOImpl(true);
+            SuitDAOImpl suitDAO = new SuitDAOImpl(true);
 
-        TransactionManager transaction = new TransactionManager(suitDAO);
-        transaction.beginTransaction();
+            TransactionManager transaction = new TransactionManager(suitDAO);
+            transaction.beginTransaction();
 
-        try {
-            Suit createdSuit = suitDAO.create(suit);
+            try {
+                Suit createdSuit = suitDAO.create(suit);
 
-            transaction.commit();
-            LOGGER.log(Level.DEBUG, "Created suit: " + createdSuit);
+                transaction.commit();
+                LOGGER.log(Level.DEBUG, "Created suit: " + createdSuit);
 
-            messages.add("Suit created successfully");
+                messages.add("Suit created successfully");
 
-            content.insertJsonAttribute("suit", createdSuit);
-            content.insertJsonAttribute("messages", messages);
-            content.insertJsonAttribute("success", true);
-        } catch (DAOException e) {
-            transaction.rollback();
-            throw new ReceiverException("Database Error: " + e.getMessage(), e);
-        } finally {
-            transaction.endTransaction();
+                content.insertJsonAttribute("suit", createdSuit);
+                content.insertJsonAttribute("messages", messages);
+                content.insertJsonAttribute("success", true);
+            } catch (DAOException e) {
+                transaction.rollback();
+
+                messages.add("Can't create suit.");
+                content.insertJsonAttribute("errors", messages);
+                content.insertJsonAttribute("success", false);
+
+                throw new ReceiverException("Database Error: " + e.getMessage(), e);
+            } finally {
+                transaction.endTransaction();
+            }
+        } else {
+            content.insertJsonAttribute("errors", validator.getErrors());
+            content.insertJsonAttribute("success", false);
         }
     }
 
@@ -104,32 +123,33 @@ public class SuitReceiverImpl extends AbstractReceiver implements SuitReceiver {
     @Override
     public void updateSuit(RequestContent content) throws ReceiverException {
         ArrayList<String> messages = new ArrayList<>();
+        SuitValidator validator = new SuitValidator();
 
-        int id = Integer.parseInt(content.findParameter("id"));
+        String idNumber = content.findParameter("id");
         String name = content.findParameter("name");
 
-        //TODO Create validators
+        if(validator.validateSuit(idNumber, name)) {
+            final int id = Integer.parseInt(idNumber);
 
-        Suit suit = new Suit(id);
-        suit.setName(name);
+            Suit suit = new Suit(id);
+            suit.setName(name);
 
-        try (SuitDAOImpl suitDAO = new SuitDAOImpl(false)) {
-            boolean result = suitDAO.update(suit);
+            try (SuitDAOImpl suitDAO = new SuitDAOImpl(false)) {
+                suitDAO.update(suit);
 
-            if (result) {
                 messages.add("Suit updated successfully");
-            } else {
+                content.insertJsonAttribute("messages", messages);
+                content.insertJsonAttribute("success", true);
+            } catch (DAOException e) {
                 messages.add("Can't update current suit");
+                content.insertJsonAttribute("errors", messages);
+                content.insertJsonAttribute("success", false);
+
+                throw new ReceiverException("Database Error: " + e.getMessage(), e);
             }
-
-            content.insertJsonAttribute("messages", messages);
-            content.insertJsonAttribute("success", result);
-        } catch (DAOException e) {
-            messages.add("Can't update current suit");
-            content.insertJsonAttribute("messages", messages);
+        } else {
+            content.insertJsonAttribute("errors", validator.getErrors());
             content.insertJsonAttribute("success", false);
-
-            throw new ReceiverException("Database Error: " + e.getMessage(), e);
         }
     }
 
@@ -141,31 +161,33 @@ public class SuitReceiverImpl extends AbstractReceiver implements SuitReceiver {
      */
     @Override
     public void removeSuit(RequestContent content) throws ReceiverException {
-        int id = Integer.parseInt(content.findParameter("id"));
         ArrayList<String> messages = new ArrayList<>();
+        SuitValidator validator = new SuitValidator();
 
-        //TODO Create validators
+        String idNumber = content.findParameter("id");
 
-        Suit suit = new Suit(id);
-        LOGGER.log(Level.DEBUG, "Want remove suit: " + suit);
+        if (validator.validateId(idNumber)) {
+            final int id = Integer.parseInt(idNumber);
+            Suit suit = new Suit(id);
 
-        try (SuitDAOImpl suitDAO = new SuitDAOImpl(false)) {
-            boolean result = suitDAO.remove(suit);
+            LOGGER.log(Level.DEBUG, "Want remove suit: " + suit);
 
-            if (result) {
+            try (SuitDAOImpl suitDAO = new SuitDAOImpl(false)) {
+                suitDAO.remove(suit);
+
                 messages.add("Suit removed successfully");
-            } else {
+                content.insertJsonAttribute("messages", messages);
+                content.insertJsonAttribute("success", true);
+            } catch (DAOException e) {
                 messages.add("Can't remove current suit");
+                content.insertJsonAttribute("errors", messages);
+                content.insertJsonAttribute("success", false);
+
+                throw new ReceiverException("Database Error: " + e.getMessage(), e);
             }
-
-            content.insertJsonAttribute("messages", messages);
-            content.insertJsonAttribute("success", result);
-        } catch (DAOException e) {
-            messages.add("Can't remove current suit");
-            content.insertJsonAttribute("messages", messages);
+        } else {
+            content.insertJsonAttribute("errors", validator.getErrors());
             content.insertJsonAttribute("success", false);
-
-            throw new ReceiverException("Database Error: " + e.getMessage(), e);
         }
     }
 }
